To solve the issue #281
The Pdf files save text data in a "Boxes" style, where any box has information like the coordinates (X,Y) where the box is and if direction where the text flow. Knowing this, you could know that a text that has the same Y value (or near) could be in the line, same with the X value. I am sayin the same or near, because the (X,Y) value will change from one Pdf to another of the same kind of pdf.

When you used the getText() function from the Smalot/PdfParser, the text came disordered. And some times is really difficult to understand which is the info you are trying to get. But, using the X,Y coordinates of the text box, help to find the right info. That information is getted internally by the Smallot/PdfParser packages, but is not public so, you can not really used it.

I used to parse PDF file in Python using a module called pdfminer or pdfminer3k. In that modules, you can get the text of the PDF with the text boxes info, that helps you in the parse process.

I am changing smalot/pdfparser package to imitate that behavior, but I need to change a little bit (make a property from protected to public), and I am using internal status of the object that could be changed in the future, so I rather ask here for change so EveryOne can use it, and that changes in internal data and status, will not affect it.

To get the information we need, for this moment, a change to the Element.php file should made. The change is that the value property should be change from protected to public. If you do that you can have something like:
```
$parser = new Parser();
$pdf = $parser->parseFile(pdfFile);
$pages = $pdf->getPages();
$page = $pages[pageNum];
$text = $page->getText();  #This is important, eventhought you don't use $text after this
$content = $page->get("Contents");
$values = $content->value;
$contentArray = [];
foreach ($values as $value) {
    $contentArray[] = $value->getContent();
}
foreach ($contentArray as $contentItems) {
    $arrayText = explode("\n", $contentItems); # split in lines to better process of the data
         foreach ($arrayText as $data) {
              $end = substr($data, strlen($data) - 2);
              if ($end == "Tm" or $end == "Tj" ) {
                   <Process Lines>
              }
        }
}
```
In <Process Lines> you do whatever you need to do, but see that you only do it with to types
of lines. The lines that ends with "Tj" and the lines that ends with "Tm". Examples are:
`1 0 0 1 365.52 755.28 Tm`
`(Some Text) Tj`
In the Tj lines, Text inside Parenthesis is the Data Text you need to work with.
In the Tm lines, there are information about the text that are in a text box. So a Tm
is presented before Tj lines and group Tj lines. The last 2 float numbers are the X and the Y of the text box coordinates, being the (0,0) the botton left corner of the Pdf page, As I explained before, same (or near) X, Y in different Pdf files of the same kind, could be treated as the "Same" text box (and the same for lines or columns).
I dont know what are the meaning of the first 2 int numbers, I couldnt find an explanation, but as far as I now if the text inside the box goes from left to right, the fourth int number has a value of 1. If the text inside the box goes up vertically, the third int number has a value of -1.

This helped to know WHERE is the Data Text in the pdf. And helps parsing the data text inside the pdf page/file.

After the change on this package, calling this:
`$extractedData = $page->extractData();`
and doing a var_dump over $extractedData, you will have something like:
```
array (size=3)
 0 =>
    array (size=3)
      0 => string '1 0 0 1 558.48 763.92' (length=21)
      1 => string 'Some Text' (length=9)
      2 => string ' ' (length=1)
 1 =>
    array (size=3)
      0 => string '1 0 0 1 577.92 761.28' (length=21)
      1 => string 'Other Text' (length=10)
      2 => string ' ' (length=1)
 2 =>
    array (size=4)
      0 => string '1 0 0 1 365.52 755.28' (length=21)
      1 => string 'More Text' (length=9)
      2 => string 'and more...' (length=11)
      3 => string ' ' (length=1)
```