Various Changes:  for properly worked
After start with the test units, there were things to be changed so the method could work with other files.  So for this the following methods were added:
extractRawData, extractDecodedRawData, getDataCommands, getDataTm and getTextXY.
The user should use getDataTm to get the Text Matrix  of every text and have a look of where to find a given text. Then he can use getTextXY to get the same info for diferent files of the same type.
A new Property was included, named $dataTm, here is saved the TextMatrix information of the given page, so others calls to the getTextXY calls do not have to get all the info again.
2 conts where added: xTm and yTm it has the index of the x and y values in the Tm data.
For the method extractDecodeRawData, a change in the method decodeContent of the Font.php file, was done, it has to be public instead of protected.
Finally, the test units tests where added to the Page class