RawDataParser: made functions protected; replaced xrefCache workaround with $xref as separate parameter
TCPDF_PARSER used "$this->xref", which made "xref" to some
kind of a global variable in the class (same for $this->pdfdata, which
was already changed earlier).

xref (and pdfdata) as class properties requires us to make sure:

- its always up-to-date and
- not altered on the way

when using it in a function. Using it as parameter instead makes its scope
more clear and related code more decoupled.