Support metadata element names containing spaces (#612)
* Update Element.php

Add ability for PdfParser to parse metadata names with hexadecimal encoded characters such as "Document#20Type" where \#20 is a space.
Resolves Issue #529

* Update ElementTest.php

Add test for spaces in metadata property names.

* Make sure we fully support hex

Too quick on the commit! Make sure our two 'digit' regexp also finds A-F hex digits. Add a test for #2d which is a hyphen.

* fixed coding style issue in Element.php

---------

Co-authored-by: Konrad Abicht <hi@inspirito.de>