Modify some code comments + q and Q update
Modify several code comments to be clearer.

Remove the `$key => ` from `decodeText()` in **Font.php** as it's no longer needed.

Also, now that `cleanContent()` is ignoring non `text/plain` content, there should be no errant `q` or `Q` commands that cause the stored-state stack to try restoring a state that doesn't exist. Remove the kludgy code that prevented this.