Remove textMatrix from decodeText()
As it turns out, the current text matrix *can* be used to position text in decodeText(), but it is actually applied **both** to the $offset and the user-selected font space limit. In this way it cancels out, and importing the text matrix into the function is not necessary. Remove two tests in FontTest.php that were associated with this, as they're no longer valid.

In decodeText() account for strings of text that have spaces with large offsets, such as
'a    line    of    text    that    is    full-width    justified.'
Treat these as single spaces instead of honouring the offsets given and spacing them out unnecessarily.

Save and use the current font-size when guessing the "fudge" factor for text block width. Make sure this is saved as well when the q command is seen.

Adjust the position factor threshold up from 10 to 24 for deciding when to insert tabs or spaces.

Also update the current horizontal position when storing the last-written-position for use by an ActualText.

Add a test for Microsoft Print-to-PDF v1.7.

This change actually reverts the output value in a ParserTest assertion back to the original value before this PR. So that's good!