More accounting for horizontal font factor
Account for the entire font-factor (font-size multiplied by the horizontal scaling factor of the text-matrix) when estimating the width of the current text block.

Insert a fix when decoding octal strings in Font::decodeOctal(), check further ahead for escaped backslashes.

Remove tests for images in DocumentGeneratorFocusTest.php. These also fail in the current v2.7.0 release and they should be looked at in a separate PR.