Prevent zero from being passed to array_chunk() (#686)
Passing a zero (0) value to the `array_chunk()` function causes an error, and in rare cases, a PDF XRef object may be added to a document with an "empty" `/W [0 0 0]` command. This would cause the `$rowlen` variable to be set to zero and cause an error.

Add a simple check to return an empty array in this case.