Add PCRE dotall modifier
Add the /s modifier so the `.` token matches newlines as well. Thanks to @iGrog for supplying another PDF that demonstrated this issue. Add the same modifier for dictionaries as well, fixing this oversight.

Move the inline image replacement before string replacement. Parentheses in binary image data may be interpreted as the start of a string.

Move the inline images test to its own function and add a newline to the sample data to test for the dotall modifier change.