Account for inline images in formatContent() (#693)
* Account for inline image data in formatContent()

`formatContent()` now accounts for inline image `BI ... ID ... EI` commands in document streams.

* Include BI command in the regexp

Include the `BI` command in the regexp, and move inline image detection after string replacement to prevent false-positives.

* Add PCRE dotall modifier

Add the /s modifier so the `.` token matches newlines as well. Thanks to @iGrog for supplying another PDF that demonstrated this issue. Add the same modifier for dictionaries as well, fixing this oversight.

Move the inline image replacement before string replacement. Parentheses in binary image data may be interpreted as the start of a string.

Move the inline images test to its own function and add a newline to the sample data to test for the dotall modifier change.

* More robust check for BI within strings

`BI` "commands" within strings should not be parsed as the beginning of inline image blocks. Detect if the `BI` we found is inside a (string) and if it is, note the offset and move past it for the next match.

* Bump search offset if no Height or Width found

In the case where a valid inline image dictionary isn't found, or if the dictionary doesn't include the required parameters Height and Width, also bump the search offset forward by the current match position so we don't fall into a loop here.

* Add "Step X:" to comments in formatContent()

Add "Step X:" to the comments to better define what the inline image replacement code is doing.

Small adjustment to the balanced parentheses regexp to also exclude open parenthesis '(' from the matching. This will ensure replacing balanced parentheses from the innermost to the outermost.

---------

Co-authored-by: Konrad Abicht <hi@inspirito.de>