Account for escaped slashes and parentheses
The regexp to detect strings only backchecked two characters for escape slashes, so if an escaped slash immediately preceded an escaped parenthesis, the script would interpret it as an escaped slash and an unescaped parenthesis.

Perform a string replace to get rid of all escaped slashes and parentheses; they aren't needed when just checking for balanced parentheses. Also add removing slashes to the inline images section above.