If no `dc:format` XMP tag, merge metadata
Previously `extractXMPMetadata()` would check for the existence of a `dc:format` tag with an `application/pdf` MIME-type value before allowing found XMP metadata to be merged with the other document details.

If the tag doesn't exist, merge the metadata anyway. If it _does_ exist, _then_ check to see if it has the `application/pdf` MIME-type.