chore: update documentation
Follow up of issue #427 using the suggested code at issue #472.

Notes:
Have PDF files that the pdfparser can't get the MediaBox, as fallback to
get the page dimension is possible to use the BBox (Bounding Box).

The Bounding Box could not have the real dimension of a page put is a
solution to have a value that could be the page dimension at cases that
isn't possible to found the MediaBox.

The BBox only was found at the first page, maybe would be good to check
if have a best way to get the page dimensions only using the MediaBox
and solving the issue that hide this information.

Signed-off-by: Vitor Mattos <vitor@php.rio>